Rename media property to dbMedia to separate handling
By changing the Hibernate mapping for the media column to the dbMedia
property, we can allow the getMedia to encapsulate the behavior of using
content hosting or the database. The new dbMedia property is public, so
it can still be used directly (by the DAO). When saving, any media bytes
set are saved to CHS and the dbMedia property is set to null. This will
remove the file from the database row. Subsequent reads will find that
field is null and look for the file in CHS. The previous interface and
behavior are preserved for client code.