Get most of the way to clear, phased property resolution
Because we allow placeholders in properties values as well as beans, we
have to be careful. If everything is in one context, the first
properties (all literal) could fill in the bean properties. But we allow
placeholders in the custom prop files, so we can't fix up beans that
might use one of the kernel keys before we load the custom files.

1. Kernel props from classpath
2. System props fill in config bean with property file definitions
3. Additional property sources, progressively expanding placeholders
4. The rest of the beans, getting the fully resolved set of values